Bug made as a component
Changed Bug to be a separate jsx component instead of a creating a table row everytime
